backend(v0.2): authz, measurement CRUD, cost aggregation, projection km_at_reference, service reminders
- Authz: delete car/tire-set admin-only; notification-log + rule CRUD scoped to car owners.
- Tires: PATCH/DELETE single measurement; Decimal->float on outputs (fixes FE NaN/blank
  pressure); projection adds km_at_reference + daily-km fallback from the set's own
  measurements (consistent projection/date).
- Service: 6 next-term columns (model + Alembic 0002); create_reminder idempotently upserts
  a ServiceInterval (intervals panel + smart notifications).
- Costs: expenses breakdown + dashboard monthly cost now include fuel (total_cost) and
  service (cost) using the existing english category keys; from_date/to_date on
  fuel/expenses/breakdown, group_by on fuel stats.
- Cars: GET /cars enriched (stk/pzp/kasko + overdue); POST /cars name optional (derived).
- Tests: new test_v02_backend.py (10). Docs (API/DB/spec) updated.

+++ backend/app/schemas/expense.py
@@ -30,12 +30,14 @@ class ExpenseUpdate(BaseModel):
 
 
 class ExpenseOut(BaseModel):
+    # ``amount`` is serialized as a JSON number (float), not Decimal-as-string,
+    # so the FE can compute on it without NaN.
     model_config = ConfigDict(from_attributes=True)
 
     id: int
     occurred_at: date
     description: str | None = None
-    amount: Decimal
+    amount: float
     category: str
